refactor damage dice roller to use sandbox renderer
